requirements.txt
numpy==1.19.5
oauthlib==3.2.2
olefile 
open-clip-torch==2.19.0
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
partd 
patsy 
Pillow==9.4.0
progress 
prompt-toolkit 
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyassimp==5.2.5
Pygments 
pymodbus==2.5.3
pyparsing 
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyserial 
PySide2==5.15.2.1
python-dateutil 
python-qt-binding==0.4.4
pytz 
PyWavelets==1.4.1
PyYAML 
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rich 
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.15
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.15
rosmake==1.15.8
rosmaster==1.15.15
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.15.15
rospkg 
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.20
scikit-image 
scikit-learn 
scipy==1.10.0
seaborn 
sensor-msgs==1.13.1
sentencepiece==0.1.99
shellingham 
shiboken2==5.15.2.1
six==1.15.0
smach==2.5.0
smach-ros==2.5.1
smclib==1.8.6
soupsieve==2.4
srdfdom==0.6.4
statsmodels 
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-gpu==2.6.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.6.13
toolz 
topic-tools==1.15.15
torch 
torchvision 
tornado 
tqdm==4.65.0
typer 
typing-extensions==3.7.4.3
tzdata==2023.3
urdfdom-py==0.4.6
urllib3==1.26.15
wcwidth 
websocket-client==0.57.0
Werkzeug==2.3.7
wrapt==1.12.1
xacro==1.14.16
yarl 
zipp==3.16.2
